+++ core/src/main/java/com/shopflix/core/data/form/ProductVariantForm.java
@@ -6,7 +6,7 @@ public class ProductVariantForm {
     public Long id;
     public String name;
     public BigDecimal price;
-    public Long quantity;
+    public Integer quantity;
     public String sku;
     public Integer position;
     public String value1;
@@ -38,12 +38,12 @@ public class ProductVariantForm {
         this.price = price;
     }
 
-    public Long getQuantity()
+    public Integer getQuantity()
     {
         return quantity;
     }
 
-    public void setQuantity(Long quantity)
+    public void setQuantity(Integer quantity)
     {
         this.quantity = quantity;
     }
